requirements.txt
selenium
webdriver-manager
colorama
flask
flask-cors
werkzeug
pymongo[srv]
python-dotenv
gunicorn
certifi
pytesseract
Pillow
cryptography
APScheduler
requests